avoid cyclic globe deps, fix unit tests
